系統/requirements.txt
streamlit
pandas
beautifulsoup4
requests
schedule
sqlalchemy
plotly
lxml
feedparser
wordcloud
duckduckgo-search
psycopg2-binary
python-dotenv
